feat:better tab bar toggeling
