--- private_dot_config/nvim/after/plugin/bufferline.lua
@@ -1,5 +1,0 @@
--- require('bufferline').setup({
---   options = {
---     mode = 'buffers',
---   }
--- })
