--- private_dot_config/nvim/after/plugin/indentline.lua
@@ -1,18 +1,0 @@
-require('ibl').setup({
-  indent = { char = "│" },
-  exclude = {
-    filetypes = {
-      "help",
-      "terminal",
-      "lazy",
-      "lspinfo",
-      "TelescopePrompt",
-      "TelescopeResults",
-      "mason",
-      "nvdash",
-      "nvcheatsheet",
-      "packer",
-    }
-  },
-  scope = { enabled = false }
-})
