--- private_dot_config/nvim/after/plugin/lspconfig.lua
@@ -1,9 +1,0 @@
-require('lspconfig').lua_ls.setup({
-  settings = {
-    Lua = {
-      diagnostics = {
-        globals = { 'vim' }
-      }
-    }
-  }
-})
