--- private_dot_config/nvim/after/plugin/null-ls.lua
@@ -1,7 +1,0 @@
-local null_ls = require('null-ls')
-
-null_ls.setup({
-  sources = {
-    null_ls.builtins.formatting.black
-  }
-})
