--- private_dot_config/nvim/after/plugin/nvim-tree.lua
@@ -1,7 +1,0 @@
--- empty setup using defaults
-require("nvim-tree").setup({
-  filters = {
-    custom = { "^.git$", "^.angular" },
-    dotfiles = true
-  }
-})
