--- private_dot_config/nvim/after/plugin/statusline.lua
@@ -1,15 +1,0 @@
-local lualine = require('lualine')
-
-
-lualine.setup({
-  options = {
-    component_separators = {
-      left = '',
-      right = ''
-    },
-    section_separators = {
-      left = '',
-      right = ''
-    }
-  }
-})
