--- private_dot_config/nvim/after/plugin/theme.lua
@@ -1,5 +1,0 @@
-require('onedark').setup({
-  style = "warmer"
-})
-
-require('onedark').load()
