--- private_dot_config/nvim/after/plugin/toggleterm.lua
@@ -1,10 +1,0 @@
-local terminal = require("toggleterm")
-
-terminal.setup({
-  size = function()
-    return vim.o.columns / 2
-  end,
-  start_in_insert = true,
-  direction = "vertical",
-  open_mapping = "<Bslash>t",
-})
