--- private_dot_config/nvim/after/plugin/zenmode.lua
@@ -1,5 +1,0 @@
-require('zen-mode').setup({
-  window = {
-    backdrop = 0.5
-  }
-})
