--- private_dot_config/nvim/ginit.vim
@@ -1,4 +1,0 @@
-GuiFont! JetBrains\ Mono:h14
-
-GuiTabline 0
-GuiPopupmenu 0
