+++ private_dot_config/nvim/init.lua
@@ -14,6 +14,9 @@ if not vim.loop.fs_stat(lazypath) then
   })
 end
 
+
 vim.opt.rtp:prepend(lazypath)
+
+require('lazy').setup("plugins")
 
 require('chaitanya')
