+++ private_dot_config/nvim/lua/chaitanya/init.lua
@@ -1,13 +1,4 @@
-require('chaitanya.plugins')
 require('chaitanya.options')
 require('chaitanya.remap')
-require('chaitanya.nvim-qt')
-require('chaitanya.gui')
-
--- require('neodev').setup()
-
--- vim.cmd [[ colorscheme gruvbox ]]
-vim.cmd("colorscheme onedark")
-
 
 vim.diagnostic.config({ virtual_text = false })
