+++ private_dot_config/nvim/lua/plugins/autoclose.lua
@@ -1,0 +1,6 @@
+return {
+  'm4xshen/autoclose.nvim',
+  config = function()
+    require('autoclose').setup()
+  end
+}
