+++ private_dot_config/nvim/lua/plugins/catppuccin.lua
@@ -1,0 +1,8 @@
+return {
+  "catppuccin/nvim",
+  name = "catppuccin",
+  priority = 1000,
+  config = function()
+    vim.cmd.colorscheme("catppuccin-frappe")
+  end
+}
