+++ private_dot_config/nvim/lua/plugins/comment.lua
@@ -1,0 +1,6 @@
+return {
+  'numToStr/Comment.nvim',
+  config = function()
+    require('Comment').setup()
+  end
+}
