+++ private_dot_config/nvim/lua/plugins/gitsigns.lua
@@ -1,0 +1,3 @@
+return {
+  'lewis6991/gitsigns.nvim',
+}
