+++ private_dot_config/nvim/lua/plugins/hlchunk.lua
@@ -1,0 +1,7 @@
+return {
+  "shellRaining/hlchunk.nvim",
+  event = { "UIEnter" },
+  opts = {
+    blank = { enable = false }
+  }
+}
