+++ private_dot_config/nvim/lua/plugins/indent-blankline.lua
@@ -1,0 +1,22 @@
+return {
+  'lukas-reineke/indent-blankline.nvim',
+  main = "ibl",
+  opts = {
+    indent = { char = "│" },
+    exclude = {
+      filetypes = {
+        "help",
+        "terminal",
+        "lazy",
+        "lspinfo",
+        "TelescopePrompt",
+        "TelescopeResults",
+        "mason",
+        "nvdash",
+        "nvcheatsheet",
+        "packer",
+      }
+    },
+    scope = { enabled = false }
+  }
+}
