+++ private_dot_config/nvim/lua/plugins/lsp-utils.lua
@@ -1,0 +1,18 @@
+return {
+  -- LSP Support
+  'neovim/nvim-lspconfig',
+  'williamboman/mason.nvim',
+  'williamboman/mason-lspconfig.nvim',
+
+  -- Autocompletion
+  'hrsh7th/nvim-cmp',
+  'hrsh7th/cmp-nvim-lsp',
+  'L3MON4D3/LuaSnip',
+
+  'rafamadriz/friendly-snippets',
+  'saadparwaiz1/cmp_luasnip',
+  'folke/neodev.nvim',
+  'hrsh7th/cmp-buffer',
+  'HiPhish/rainbow-delimiters.nvim',
+  'stevearc/dressing.nvim',
+}
