+++ private_dot_config/nvim/lua/plugins/lspkind.lua
@@ -1,0 +1,6 @@
+return {
+  'onsails/lspkind.nvim',
+  config = function() 
+    require('lspkind').init({ preset = 'codicons' })
+  end
+}
