+++ private_dot_config/nvim/lua/plugins/lualine.lua
@@ -1,0 +1,10 @@
+return {
+  'nvim-lualine/lualine.nvim',
+  dependencies = { 'kyazdani42/nvim-web-devicons', opt = true },
+  opts = {
+    options = {
+      section_separators = { left = '', right = '' },
+      component_separators = { left = '', right = '' }
+    }
+  }
+}
