+++ private_dot_config/nvim/lua/plugins/noegit.lua
@@ -1,0 +1,10 @@
+return {
+  'NeogitOrg/neogit',
+  dependencies = {
+    'nvim-lua/plenary.nvim',           -- required
+    'nvim-telescope/telescope.nvim',   -- optional
+    'sindrets/diffview.nvim',          -- optional
+    'ibhagwan/fzf-lua',                -- optional
+  },
+  config = true
+}
