+++ private_dot_config/nvim/lua/plugins/notify.lua
@@ -1,0 +1,4 @@
+return {
+  'rcarriga/nvim-notify',
+  config = true
+}
