+++ private_dot_config/nvim/lua/plugins/nvim-surround.lua
@@ -1,0 +1,5 @@
+return {
+  'kylechui/nvim-surround',
+  version = '*',
+  config = function() require('nvim-surround').setup() end
+}
