+++ private_dot_config/nvim/lua/plugins/nvim-tree.lua
@@ -1,0 +1,9 @@
+return {
+  'nvim-tree/nvim-tree.lua',
+  opts = {
+    filters = {
+      custom = { "^.git$", "^.angular" },
+      dotfiles = true
+    }
+  }
+}
