+++ private_dot_config/nvim/lua/plugins/onedark.lua
@@ -1,0 +1,6 @@
+return {
+  'navarasu/onedark.nvim',
+  config = function()
+    require('onedark').setup()
+  end
+}
