+++ private_dot_config/nvim/lua/plugins/zen-mode.lua
@@ -1,0 +1,12 @@
+return {
+  'folke/zen-mode.nvim',
+  opts = {
+    window = {
+      backdrop = 0.9,
+      options = {
+        number = true,
+        cursorline = true,
+      }
+    }
+  }
+}
