--- private_dot_config/wezterm/settings.json
@@ -1,1 +1,0 @@
-{"tab_bar_is_visible":true}
